requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
pydantic
pandas
tabulate
python-multipart
psycopg2